+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryEffect.kt
@@ -53,6 +53,8 @@ data class CheckPatientReassignmentStatus(
 
 data class MarkDiabetesDiagnosis(val patientUuid: UUID) : PatientSummaryEffect()
 
+data class MarkHypertensionDiagnosis(val patientUuid: UUID) : PatientSummaryEffect()
+
 sealed class PatientSummaryViewEffect : PatientSummaryEffect()
 
 data class HandleEditClick(
@@ -109,3 +111,5 @@ data class ShowReassignPatientWarningSheet(
 ) : PatientSummaryViewEffect()
 
 data object ShowDiabetesDiagnosisWarning : PatientSummaryViewEffect()
+
+data class ShowHypertensionDiagnosisWarning(val continueToDiabetesDiagnosisWarning: Boolean) : PatientSummaryViewEffect()