+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryUiActions.kt
@@ -38,4 +38,5 @@ interface PatientSummaryUiActions {
       sheetOpenedFrom: ReassignPatientSheetOpenedFrom
   )
   fun showDiabetesDiagnosisWarning()
+  fun showHypertensionDiagnosisWarning(continueToDiabetesDiagnosisWarning: Boolean)
 }