+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryUpdate.kt
@@ -86,11 +86,38 @@ class PatientSummaryUpdate(
       is PatientReassignmentWarningClosed -> patientReassignmentWarningClosed(model, event)
       HasDiabetesClicked -> dispatch(MarkDiabetesDiagnosis(model.patientUuid))
       is DiagnosisWarningResultReceived -> diagnosisWarningResultReceived(event.diagnosisWarningResult)
+      is HasHypertensionClicked -> hasHypertensionClicked(event.continueToDiabetesDiagnosisWarning, model.patientUuid)
+      is HypertensionNotNowClicked -> hypertensionNotNowClicked(event.continueToDiabetesDiagnosisWarning)
+    }
+  }
+
+  private fun hypertensionNotNowClicked(continueToDiabetesDiagnosisWarning: Boolean): Next<PatientSummaryModel, PatientSummaryEffect> {
+    return if (continueToDiabetesDiagnosisWarning) {
+      dispatch(ShowDiabetesDiagnosisWarning)
+    } else {
+      noChange()
+    }
+  }
+
+  private fun hasHypertensionClicked(
+      continueToDiabetesDiagnosisWarning: Boolean,
+      patientUuid: UUID
+  ): Next<PatientSummaryModel, PatientSummaryEffect> {
+    return if (continueToDiabetesDiagnosisWarning) {
+      dispatch(MarkHypertensionDiagnosis(patientUuid), ShowDiabetesDiagnosisWarning)
+    } else {
+      dispatch(MarkHypertensionDiagnosis(patientUuid))
     }
   }
 
   private fun diagnosisWarningResultReceived(diagnosisWarningResult: DiagnosisWarningResult): Next<PatientSummaryModel, PatientSummaryEffect> {
-    return dispatch(ShowDiabetesDiagnosisWarning)
+    val effect = when (diagnosisWarningResult) {
+      DiagnosisWarningResult.DiabetesWarning -> ShowDiabetesDiagnosisWarning
+      DiagnosisWarningResult.HypertensionWarning -> ShowHypertensionDiagnosisWarning(continueToDiabetesDiagnosisWarning = false)
+      DiagnosisWarningResult.BothDiagnosisWarning -> ShowHypertensionDiagnosisWarning(continueToDiabetesDiagnosisWarning = true)
+    }
+
+    return dispatch(effect)
   }
 
   private fun patientReassignmentWarningClosed(