+++ app/src/test/java/org/simple/clinic/summary/PatientSummaryUpdateTest.kt
@@ -1777,6 +1777,72 @@ class PatientSummaryUpdateTest {
         ))
   }
 
+  @Test
+  fun `when diagnosis warning result is htn warning, show htn warning dialog`() {
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(DiagnosisWarningResultReceived(DiagnosisWarningResult.HypertensionWarning))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(ShowHypertensionDiagnosisWarning(continueToDiabetesDiagnosisWarning = false))
+        ))
+  }
+
+  @Test
+  fun `when diagnosis warning result is both warnings, show htn warning dialog with continue to diabetes warning`() {
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(DiagnosisWarningResultReceived(DiagnosisWarningResult.BothDiagnosisWarning))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(ShowHypertensionDiagnosisWarning(continueToDiabetesDiagnosisWarning = true))
+        ))
+  }
+
+  @Test
+  fun `when has hypertension is clicked and doesn't have diabetes warning, then mark the hypertension diagnosis`() {
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(HasHypertensionClicked(continueToDiabetesDiagnosisWarning = false))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(MarkHypertensionDiagnosis(patientUuid))
+        ))
+  }
+
+  @Test
+  fun `when has hypertension is clicked and also have diabetes warning, then mark the hypertension diagnosis and show diabetes diagnosis warning`() {
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(HasHypertensionClicked(continueToDiabetesDiagnosisWarning = true))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(MarkHypertensionDiagnosis(patientUuid), ShowDiabetesDiagnosisWarning)
+        ))
+  }
+
+  @Test
+  fun `when hypertension diagnosis not now is clicked and also have diabetes warning, then show diabetes diagnosis warning`() {
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(HypertensionNotNowClicked(continueToDiabetesDiagnosisWarning = true))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(ShowDiabetesDiagnosisWarning)
+        ))
+  }
+
+  @Test
+  fun `when hypertension diagnosis not now is clicked and doesn't have diabetes warning, then do nothing`() {
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(HypertensionNotNowClicked(continueToDiabetesDiagnosisWarning = false))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasNoEffects()
+        ))
+  }
+
   private fun PatientSummaryModel.forExistingPatient(): PatientSummaryModel {
     return copy(openIntention = ViewExistingPatient)
   }